+++ poem-openapi/src/payload/attachment.rs
@@ -1,15 +1,41 @@
-use poem::{http::HeaderValue, Body, IntoResponse, Response};
+use std::fmt::Write;
+
+use poem::{http::header::CONTENT_DISPOSITION, Body, IntoResponse, Response};
 
 use crate::{
     payload::{Binary, Payload},
-    registry::{MetaResponses, MetaSchemaRef, Registry},
+    registry::{MetaHeader, MetaMediaType, MetaResponse, MetaResponses, MetaSchemaRef, Registry},
+    types::Type,
     ApiResponse,
 };
+
+const CONTENT_DISPOSITION_DESC: &str = "Indicate if the content is expected to be displayed inline in the browser, that is, as a Web page or as part of a Web page, or as an attachment, that is downloaded and saved locally.";
+
+/// Attachment type
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub enum AttachmentType {
+    /// Indicate it can be displayed inside the Web page, or as the Web page
+    Inline,
+    /// Indicate it should be downloaded; most browsers presenting a 'Save as'
+    /// dialog
+    Attachment,
+}
+
+impl AttachmentType {
+    #[inline]
+    fn as_str(&self) -> &'static str {
+        match self {
+            AttachmentType::Inline => "inline",
+            AttachmentType::Attachment => "attachment",
+        }
+    }
+}
 
 /// A binary payload for download file.
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct Attachment<T> {
     data: Binary<T>,
+    ty: AttachmentType,
     filename: Option<String>,
 }
 
@@ -18,8 +44,15 @@ impl<T: Into<Body> + Send> Attachment<T> {
     pub fn new(data: T) -> Self {
         Self {
             data: Binary(data),
+            ty: AttachmentType::Inline,
             filename: None,
         }
+    }
+
+    /// Specify the attachment. (defaults to: [`AttachmentType::Inline`])
+    #[must_use]
+    pub fn attachment_type(self, ty: AttachmentType) -> Self {
+        Self { ty, ..self }
     }
 
     /// Specify the file name.
@@ -29,6 +62,22 @@ impl<T: Into<Body> + Send> Attachment<T> {
             filename: Some(filename.into()),
             ..self
         }
+    }
+
+    fn content_disposition(&self) -> String {
+        let mut content_disposition = self.ty.as_str().to_string();
+
+        if let Some(legal_filename) = self.filename.as_ref().map(|filename| {
+            filename
+                .replace('\\', "\\\\")
+                .replace('\"', "\\\"")
+                .replace('\r', "\\\r")
+                .replace('\n', "\\\n")
+        }) {
+            _ = write!(content_disposition, "; filename=\"{}\"", legal_filename);
+        }
+
+        content_disposition
     }
 }
 
@@ -42,27 +91,32 @@ impl<T: Into<Body> + Send> Payload for Attachment<T> {
 
 impl<T: Into<Body> + Send> IntoResponse for Attachment<T> {
     fn into_response(self) -> Response {
-        let mut resp = self.data.into_response();
-
-        if let Some(header_value) = self.filename.as_ref().and_then(|filename| {
-            let legal_filename = filename
-                .replace('\\', "\\\\")
-                .replace('\"', "\\\"")
-                .replace('\r', "\\\r")
-                .replace('\n', "\\\n");
-            HeaderValue::from_str(&format!("attachment; filename=\"{}\"", legal_filename)).ok()
-        }) {
-            resp.headers_mut()
-                .insert("Content-Disposition", header_value);
-        }
-
-        resp
+        let content_disposition = self.content_disposition();
+        self.data
+            .with_header(CONTENT_DISPOSITION, content_disposition)
+            .into_response()
     }
 }
 
 impl<T: Into<Body> + Send> ApiResponse for Attachment<T> {
     fn meta() -> MetaResponses {
-        Binary::<T>::meta()
+        MetaResponses {
+            responses: vec![MetaResponse {
+                description: "",
+                status: Some(200),
+                content: vec![MetaMediaType {
+                    content_type: Self::CONTENT_TYPE,
+                    schema: Self::schema_ref(),
+                }],
+                headers: vec![MetaHeader {
+                    name: "Content-Disposition".to_string(),
+                    description: Some(CONTENT_DISPOSITION_DESC.to_string()),
+                    required: true,
+                    deprecated: false,
+                    schema: String::schema_ref(),
+                }],
+            }],
+        }
     }
 
     fn register(_registry: &mut Registry) {}